+++ src/components/pages/articles/common/types.ts
@@ -15,3 +15,20 @@ export interface IdProps {
   articleId: number;
   categoryId: number;
 }
+
+export type WritingInputState = {
+  title: string;
+  content: string;
+  categoryId: number;
+};
+
+export enum WritingInputStateEnum {
+  CHANGE_INPUT,
+  LOAD_ARTICLE,
+}
+
+export type WritingInputStateAction = {
+  type: WritingInputStateEnum;
+  name: string;
+  value: any;
+};
